requirements.txt
numpy==1.12.1
scipy==0.18.1
tensorflow_gpu==0.12.1
toml==0.9.2
